Day 9 - Part One
- Add support for large numbers by using BigInteger
- Add support for expandable memory

+++ src/main/kotlin/intcode/Intcode.kt
@@ -8,7 +8,7 @@ class Intcode(private val memory: Memory, private val io: IOInterface = ConsoleI
         var pointer = 0
         do {
             val code = memory.getAddressValue(pointer)
-            val opCode = OpCode(pointer, code, io)
+            val opCode = OpCode(pointer, code.toInt(), io)
             val instruction = opCode.getInstruction()
             println("$parent >>> $instruction")
             instruction.execute(memory)
