+++ src/main/kotlin/intcode/params/ImmediateParameter.kt
@@ -1,13 +1,14 @@
 package intcode.params
 
 import intcode.Memory
+import java.math.BigInteger
 
 data class ImmediateParameter(val pointer: Int) : Parameter {
-    override fun getValue(memory: Memory): Int {
+    override fun getValue(memory: Memory): BigInteger {
         return memory.getAddressValue(pointer)
     }
 
-    override fun setValue(value: Int, memory: Memory) {
+    override fun setValue(value: BigInteger, memory: Memory) {
         memory.setAddressValue(pointer, value)
     }
 }
