+++ src/main/kotlin/intcode/params/Parameter.kt
@@ -1,8 +1,9 @@
 package intcode.params
 
 import intcode.Memory
+import java.math.BigInteger
 
 interface Parameter {
-    fun getValue(memory: Memory): Int
-    fun setValue(value: Int, memory: Memory)
+    fun getValue(memory: Memory): BigInteger
+    fun setValue(value: BigInteger, memory: Memory)
 }
