+++ src/test/kotlin/amplifier/AmplifierControllerSoftwareSpec.kt
@@ -5,6 +5,7 @@ import kotlinx.coroutines.runBlocking
 import org.amshove.kluent.shouldEqual
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
+import java.math.BigInteger
 
 object AmplifierControllerSoftwareSpec : Spek({
     describe("An AmplifierControllerSoftware") {
@@ -12,31 +13,31 @@ object AmplifierControllerSoftwareSpec : Spek({
             val memory = Memory("3,15,3,16,1002,16,10,16,1,16,15,15,4,15,99,0,0")
             val software = AmplifierControllerSoftware()
             val maxSignal = runBlocking { software.maxSignal(listOf(4, 3, 2, 1, 0), memory) }
-            maxSignal shouldEqual 43210
+            maxSignal shouldEqual BigInteger.valueOf(43210)
         }
         it("computes max signal #2") {
             val memory = Memory("3,23,3,24,1002,24,10,24,1002,23,-1,23,101,5,23,23,1,24,23,23,4,23,99,0,0")
             val software = AmplifierControllerSoftware()
             val maxSignal = runBlocking { software.maxSignal(listOf(0, 1, 2, 3, 4), memory) }
-            maxSignal shouldEqual 54321
+            maxSignal shouldEqual BigInteger.valueOf(54321)
         }
         it("computes max signal #2") {
             val memory = Memory("3,31,3,32,1002,32,10,32,1001,31,-2,31,1007,31,0,33,1002,33,7,33,1,33,31,31,1,32,31,31,4,31,99,0,0,0")
             val software = AmplifierControllerSoftware()
             val maxSignal = runBlocking { software.maxSignal(listOf(1, 0, 4, 3, 2), memory) }
-            maxSignal shouldEqual 65210
+            maxSignal shouldEqual BigInteger.valueOf(65210)
         }
         it("computes max signal #4") {
             val memory = Memory("3,26,1001,26,-4,26,3,27,1002,27,2,27,1,27,26,27,4,27,1001,28,-1,28,1005,28,6,99,0,0,5")
             val software = AmplifierControllerSoftware()
             val maxSignal = runBlocking { software.maxSignal(listOf(9, 8, 7, 6, 5), memory) }
-            maxSignal shouldEqual 139629729
+            maxSignal shouldEqual BigInteger.valueOf(139629729)
         }
         it("computes max signal #5") {
             val memory = Memory("3,52,1001,52,-5,52,3,53,1,52,56,54,1007,54,5,55,1005,55,26,1001,54,-5,54,1105,1,12,1,53,54,53,1008,54,0,55,1001,55,1,55,2,53,55,53,4,53,1001,56,-1,56,1005,56,6,99,0,0,0,0,10")
             val software = AmplifierControllerSoftware()
             val maxSignal = runBlocking { software.maxSignal(listOf(9, 7, 8, 5, 6), memory) }
-            maxSignal shouldEqual 18216
+            maxSignal shouldEqual BigInteger.valueOf(18216)
         }
     }
 })
