+++ src/test/kotlin/intcode/instructions/InputInstructionSpec.kt
@@ -1,7 +1,7 @@
 package intcode.instructions
 
 import intcode.Memory
-import intcode.MockIO
+import io.MockIO
 import kotlinx.coroutines.runBlocking
 import org.amshove.kluent.shouldEqual
 import org.spekframework.spek2.Spek
@@ -10,7 +10,7 @@ import org.spekframework.spek2.style.specification.describe
 object InputInstructionSpec : Spek({
     describe("An InputInstruction") {
         it("set input to memory address") {
-            val inputCode = 1123
+            val inputCode = 1123.toBigInteger()
             val memory = Memory("3,1,99")
             val inputInstruction = InputInstruction(0, MockIO(inputCode))
             runBlocking { inputInstruction.execute(memory) }
